requirements.txt
streamlit
python-dotenv
langchain
langchain_community
langchainhub
tavily-python
faiss-cpu
openai
pypdf